Add files via upload
I made Additional Homepage and Portfolio changes per Cameron's request. I added multiple columns for buying and selling stocks, recent transactions, watchlist, and portfolio summary. For the homepage, I plan to fill in the empty graphs for Nasdaq and S&P 500 when the backend is done. Top movers will be included as well.