Add files via upload
Added a controller folder and a middleware for webtokens. I also broke up app.js into seperate js files for the other webpages. I got the login and register pages to work and displayed the first and last name to the welcome message in the homepage's header. I will work on portfolio and home page back-end tomorrow.